move code into submodules to make it easier to write tests for
