+++ custom_components/keymaster/const.py
@@ -3,6 +3,7 @@ VERSION = "0.0.45"
 ISSUE_URL = "https://github.com/FutureTense/keypaster"
 PLATFORM = "sensor"
 ZWAVE_NETWORK = "zwave_network"
+MANAGER = "manager"
 
 # Attributes
 ATTR_NAME = "lockname"
